+++ inbox/constants.py
@@ -1,0 +1,1 @@
+MAX_MESSAGE_BODY_PARSE_LENGTH = 100 * 1024 * 1024  # 100 MB
