+++ inbox/models/constants.py
@@ -3,4 +3,3 @@
 MAX_INDEXABLE_LENGTH = 191
 MAX_FOLDER_NAME_LENGTH = MAX_INDEXABLE_LENGTH
 MAX_LABEL_NAME_LENGTH = MAX_INDEXABLE_LENGTH
-MAX_MESSAGE_BODY_PARSE_LENGTH = 100 * 1024 * 1024  # 100 MB
